Move Category and Tag classes to subdir
